feat: refactoring package
