+++ sanic_openapi/__init__.py
@@ -1,5 +1,7 @@
-from .oas3 import openapi
-from .swagger import oas3_blueprint, swagger_blueprint
+from .openapi3 import openapi3_blueprint, openapi
+from .openapi2 import openapi2_blueprint, doc
+
+swagger_blueprint = openapi2_blueprint
 
 __version__ = "0.6.2"
-__all__ = ["swagger_blueprint", "oas3_blueprint", "openapi"]
+__all__ = ["openapi2_blueprint", "swagger_blueprint", "openapi3_blueprint", "openapi", "doc"]
