+++ sanic_openapi/openapi2/__init__.py
@@ -1,0 +1,4 @@
+from .blueprint import blueprint_factory
+
+
+openapi2_blueprint = blueprint_factory()
