+++ sanic_openapi/openapi2/spec.py
