+++ sanic_openapi/openapi3/types.py
